# openapi/api.yaml
...
openapi: 3.0.3
info:
  title: SurenSaaS API
  version: 1.0.0
  description: API multi-PME pour construction, nettoyage et diagnostic énergétique
servers:
  - url: https://back-xxx.run.app/api/v1
    description: Production
paths:
...
  /{org}/users/me:
    get:
      summary: Profil utilisateur connecté
      security:
        - bearerAuth: []
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ============================================
  # AUTHENTICATION
  # ============================================
...
  /users/me:
    get:
      summary: Profil utilisateur connecté
      responses:
        200:
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        401:
          description: Non authentifié

  # ============================================
  # CLIENTS
  # ============================================
...
  /{org}/telegram/webhook/{bot_token}:
    post:
      summary: Webhook Telegram (public)
      description: Reçoit les updates du bot Telegram. Ne nécessite pas d'authentification car vérifié par secret.
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: bot_token
          in: path
          required: true
          schema:
            type: string
          description: Hash partiel du token pour identifier le bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Update Telegram standard
      responses:
        200:
          description: OK
        401:
          description: Secret invalide
        404:
          description: Bot non trouvé
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        org_id:
          type: string
          format: uuid
        org_slug:
          type: string
        org_name:
          type: string
        role:
          type: string
          enum: [admin, user]
        created_at:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        org_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

    # ----------------------------------------
    # NOUVEAUX SCHEMAS (AUTRES)
    # ----------------------------------------

    # Telegram
...